+++ lib/component/screen/sliver_screen.dart
@@ -200,8 +200,7 @@ class _SliverScreenState extends State<SliverScreen> {
                     sliver: Pylon<ArbitraryHeaderSpace?>(
                       value: ArbitraryHeaderSpace(headerSize),
                       local: true,
-                      builder: (context) =>
-                          widget.sliver.toSliver(context, fillRemaining: true),
+                      builder: (context) => widget.sliver,
                     )))),
         SliverToBoxAdapter(
           child: SizedBox(