+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 3.10.7
+* Additional Sliver Fixes
+* ADD `ArcaneScreen` which switches between `FillScreen` and `SliverScreen` based on the child provided
+
 ## 3.10.6
 * Sliver Fixes